chat

+++ public/script.js
@@ -96,14 +96,14 @@ logOut.addEventListener("click",()=>{
 
 
 window.onload= (function(){
-
+if (localStorage.getItem("name")){
   if (document.cookie === ""){
       localStorage.clear()
       chatHistory.innerHTML = ""
       location.href = location.origin
     }
  
- 
+  }
 })()
 
 if (!window.closed){
